Fixed bug in 64-bit amzinet
amzinet.cpp mistakenly used 'long' as an integer to cast TERMs.  'long'
is 32-bits (in C++, but 64-bits in C#).  A term is a pointer, 64-bits on
64-bit version.  This problem only manifested itself on machines that
actually needed more than 32-bits for pointers.  Truncating didn't hurt
otherwise.  The 'long's have been replaced with INTTERM, which is
64-bits in 64-bit environments, and otherwise 32-bits.

+++ samples/dotnet/cs/pets/PetsForm.cs
@@ -133,7 +133,7 @@ namespace pets
 		private void GetPetButton_Click(object sender, System.EventArgs e)
 		{
 			LogicServer ls;
-			int term;
+			long term;
 
             term = 0;
 			try
